+++ ios/CommonCharsets.swift
@@ -1,0 +1,13 @@
+//
+//  CommonCharsets.swift
+//  RNBluetoothClassic
+//
+//  Created by Ken Davidson on 2020-01-12.
+//  Copyright © 2020 Facebook. All rights reserved.
+//
+
+import Foundation
+
+enum CommonCharsets : String, CaseIterable {
+    case latin = String.Encoding.latin
+}