+++ ios/RNBluetoothClassic.m
@@ -62,6 +62,10 @@ RCT_EXTERN_METHOD(setDelimiter: (NSString)delimiter
                   resolver: (RCTPromiseResolveBlock)resolve
                   rejecter: (RCTPromiseRejectBlock)reject)
 
+RCT_EXTERN_METHOD(setEncoding: (NSInteger)code
+                  resolver: (RCTPromiseResolveBlock)resolve
+                  rejecter: (RCTPromiseRejectBlock)reject)
+
 RCT_EXTERN_METHOD(setAdapterName: (NSString)name
                   resolver: (RCTPromiseResolveBlock)resolve
                   rejecter: (RCTPromiseRejectBlock)reject)